[ares] refactor snapshots
