+++ rust/ares/src/main.rs
@@ -27,6 +27,7 @@ fn main() -> io::Result<()> {
         ares::serf::use_gdb();
         ares::serialization::use_gdb();
         ares::snapshot::use_gdb();
+        ares::snapshot::double_jam::use_gdb();
     }
 
     if filename == "serf" {
